% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdi_per_tf.R
\name{calc_bd_metrics}
\alias{calc_bd_metrics}
\title{calculate seasonal bd metrics}
\usage{
calc_bd_metrics(df, class)
}
\arguments{
\item{class}{The class for which the metrics need to be calculated}

\item{tfid}{The tiny forest id for which the metrics need to be calculated}
}
\value{
A list containing the following:

metrics  : data.frame containing monthly values of the metrics
matrix   : data.frame containing monthly count of species for each site, in wide format
plot     : a ggplot object showing monthly variation in richness
}
\description{
This function calculates various biodiversity metrics (richness and diversity) for a given tiny forest and class
}
\examples{
calc_bd_metrics(85, 'Aves')

}
