% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_species_matrix.R
\name{create_species_matrix}
\alias{create_species_matrix}
\title{Create a species matrix from a dataframe}
\usage{
create_species_matrix(df, class = NULL)
}
\arguments{
\item{df}{a dataframe with columns 'tf_id', 'species', and 'classs'}

\item{class}{a string indicating the ecological class of interest}
}
\value{
a list with two elements: sp, a species-by-tf_id matrix with species counts; sites,
a vector of unique tf_id values indicating the sites represented in the matrix.
}
\description{
This function takes an input dataframe, with species and tf_id columns, and creates a matrix
with species as the rows and tf_id as the columns. Entries in the matrix are counts of
the number of times each species occurs at each site (as represented by tf_id).
}
\examples{
data(iris)
iris_species <- iris \%>\% select(Species) \%>\% mutate(classs = 'example') \%>\% rename(species = Species, tf_id = classs)
create_species_matrix(iris_species, 'example')

}
