% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_coords.R
\name{decimal_from_deg_min}
\alias{decimal_from_deg_min}
\title{Calculate decimal coordinates from degrees/minutes}
\usage{
decimal_from_deg_min(degrees, minutes, point)
}
\arguments{
\item{point}{}
}
\value{
data frame
}
\description{
Calculate decimal coordinates from degrees/minutes
}
\examples{
decimal_from_deg_min(degrees, minutes, point)

}
