% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_from_en.R
\name{decimal_from_en}
\alias{decimal_from_en}
\title{Get decimal coordinates from eastings and northings}
\usage{
decimal_from_en(easting, northing)
}
\value{
data frame
}
\description{
Uses https://webapps.bgs.ac.uk/data/webservices/CoordConvert_LL_BNG.cfc?method=BNGtoLatLng
}
