% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nbn_buffer.R
\name{download_nbn_data}
\alias{download_nbn_data}
\title{Download species occurrence data from NBN Atlas}
\usage{
download_nbn_data(lon, lat, radius = 1L)
}
\arguments{
\item{lon}{The longitude of the location.}

\item{lat}{The latitude of the location.}

\item{radius}{The radius around the location to search for occurrences (default is 1L).}
}
\value{
A data frame containing the downloaded species occurrence data.
}
\description{
This function downloads species occurrence data from the NBN Atlas for a specified location.
}
\examples{
download_nbn_species(0, 0)

}
