% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bsbi_data.R
\name{get_bsbi_data}
\alias{get_bsbi_data}
\title{Retrieves BSBI( Botanical Society of Britain and Ireland) data for a given grid reference}
\usage{
get_bsbi_data(grid_ref)
}
\arguments{
\item{grid_ref}{A valid UK Ordnance Survey grid reference}
}
\value{
A list containing the records in the BSBI database for the given grid reference
}
\description{
Retrieves BSBI( Botanical Society of Britain and Ireland) data for a given grid reference
}
\examples{
get_bsbi_data("NY7250")
get_bsbi_data("TQ3080")
}
