% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dw_landcover.R
\name{get_dw_landcover}
\alias{get_dw_landcover}
\title{Get land cover data from the Dynamic World Land Cover dataset for a given location}
\usage{
get_dw_landcover(
  tf_id = tf_id,
  lon = -0.25083,
  lat = 52.5739214,
  dist = 1000,
  start_date = "2023-01-01",
  end_date = "2023-08-01"
)
}
\arguments{
\item{tf_id}{text: specify which Tiny Forest site ID to use, defaults to tf_id}

\item{lon}{numeric: longitude of the center of the area of interest, defaults to 0.0969}

\item{lat}{numeric: latitude of the center of the area of interest, defaults to 51.578}

\item{dist}{numeric: distance around the center point to create a circular buffer in meters, defaults to 500}

\item{start_date}{character: starting date of the image collection, defaults to '2022-01-01'}

\item{end_date}{character: ending date of the image collection, defaults to '2022-12-31'}
}
\value{
list: a list containing the following items:
\item{map}{a map object showing the image collection and area of interest}
\item{image_dates}{vector of image dates in the collection}
\item{buffer}{the buffer distance used to select the area of interest}
}
\description{
Downloads the image collection from the Dynamic World Land Cover (DW) dataset on Google Earth Engine (GEE) platform, filters it using a given location and time frame, and returns the land cover statistics for the given area.
}
\examples{
\dontrun{
# Get data for the default location
get_dw_landcover()

# Get data for a specific location and buffer distance
get_dw_landcover(tf_id= "85", lon = -1.472293, lat = 51.78398, dist = 1000)
}

}
