% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Get Elevation}
\usage{
get_elevation(lat, lon, api_key)
}
\arguments{
\item{lat}{The latitude of the location.}

\item{lon}{The longitude of the location.}

\item{api_key}{The API key required to make the API call.}
}
\value{
The elevation at the given location.
}
\description{
This function retrieves the elevation at a given longitude and latitude using the Google Maps Elevation API.
}
\examples{
get_elevation(-122.084, 37.422, 'your_api_key')
}
