% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{Get locations}
\usage{
get_locations(text)
}
\arguments{
\item{text}{}
}
\value{
data frame
}
\description{
Extracts locations from text. Uses spaCy which is a state of the art natural language processing tools and returns a data frame
}
\examples{
get_locations(text)

}
