% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phylopic.R
\name{get_phylopic_image}
\alias{get_phylopic_image}
\title{Get PhyloPic Image}
\usage{
get_phylopic_image(search, path)
}
\arguments{
\item{search}{A character string containing the name of the organism or scientific name of the taxon.}

\item{path}{A character string specifying the directory to save the image file.}
}
\value{
The image is saved to the directory specified by the path parameter.
}
\description{
This function retrieves an image from PhyloPic API and saves it to a specified directory.
}
\examples{
get_phylopic_image("Homo sapiens", "C:/images/")
get_phylopic_image("Gorilla gorilla", "D:/pictures/")
}
