% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s2_landcover.R
\name{get_s2_landcover}
\alias{get_s2_landcover}
\title{Get S2 Landcover}
\usage{
get_s2_landcover(
  tf_id = 85,
  lon = -0.25083,
  lat = 52.5739214,
  dist = 1000,
  start_date = "2023-01-01",
  end_date = "2023-09-01"
)
}
\arguments{
\item{tf_id}{Numeric ID associated with the training forest \code{default = tf_id}.}

\item{lon}{Numeric, longitude coordinate of the location \code{\link{default} = -0.250830}.}

\item{lat}{Numeric, latitude coordinate of the location \code{\link{default} = 52.5739214}.}

\item{dist}{Numeric, buffer distance in meters around the location \code{\link{default} = 1000}.}

\item{start_date}{Character string, start date of the time period in 'YYYY-MM-DD' format \code{\link{default} = '2023-01-01'}.}

\item{end_date}{Character string, end date of the time period in 'YYYY-MM-DD' format \code{\link{default} = '2023-09-01'}.}
}
\description{
Function to retrieve Sentinel-2 landcover data for a specified location and time period.
}
\examples{
## Specify the tf_id
tf_id <- 85

## Specify the location coordinates
lon <- -0.250830
lat <- 52.5739214

## Specify the buffer distance
dist <- 1000

## Specify the time period
start_date <- '2023-01-01'
end_date <- '2023-09-01'

## Call the get_s2_landcover function
get_s2_landcover(tf_id = tf_id, lon = lon, lat = lat, dist = dist)
}
