% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf_data.R
\name{get_tf_data}
\alias{get_tf_data}
\title{Get Tiny Forest Data}
\usage{
get_tf_data()
}
\arguments{
\item{url}{The URL to scrape for tiny forest data.}
}
\value{
A list containing several objects:
\describe{
\item{id_table} A data frame containing information about each tiny forest, including the URL, stub, and tf_id.
\item{tidy_tf} A tidy data frame with cleaned and transformed information about each tiny forest.
\item{tidy_tf_sf} A spatial data frame with longitude and latitude columns, suitable for mapping.
\item{tidy_tf_uk_map} A map showing the locations of the tiny forests in the UK and Ireland.
}
}
\description{
This function scrapes information about tiny forests from a specified URL and creates a data frame with the scraped information.
}
\examples{
\dontrun{
  get_tf_data()
}

}
