% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf_images.R
\name{get_tf_images}
\alias{get_tf_images}
\title{Save ggmap Images}
\usage{
get_tf_images(lon, lat, key, zoom = 15, tf_id)
}
\arguments{
\item{lon}{longitude of the center of the image}

\item{lat}{latitude of the center of the image}

\item{key}{Google API key for ggmap imagery}

\item{zoom}{degree of zoom of image; default = 15}

\item{tf_id}{earthwatch id of Tiny Forest}
}
\description{
This function saves a ggmap satellite image given a longitude and latitude.
}
\examples{
get_ggmap_images(-71.0636, 42.3581, "API_key", zoom=19)

}
