% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vegetation_indices.R
\name{get_vegetation_indices}
\alias{get_vegetation_indices}
\title{Get Vegetation Indices}
\usage{
get_vegetation_indices(
  tf_id,
  buffer = 1000,
  lat,
  lon,
  start = "2023-01-01",
  end = "2023-08-01"
)
}
\arguments{
\item{tf_id}{The ID of the Tiny Forest in the tinyForestR package.}

\item{buffer}{The buffer distance in meters around the region of interest. Default is 1000 meters.}

\item{lat}{The latitude of the center of the region of interest.}

\item{lon}{The longitude of the center of the region of interest.}

\item{start}{The start date of the image collection. Default is '2023-01-01'.}

\item{end}{The end date of the image collection. Default is '2023-08-01'.}
}
\description{
This function retrieves vegetation indices from Sentinel-2 satellite imagery within a specified region of interest.
}
