% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_features_osm.R
\name{get_water_features}
\alias{get_water_features}
\title{Get Natural Water Features}
\usage{
get_water_features(df, tf_id = 85, buffer = 1000)
}
\arguments{
\item{df}{A data.frame containing the location data.}

\item{tf_id}{An integer specifying the focal point of the analysis.}

\item{buffer}{A numeric value specifying the desired buffer size in meters.}
}
\value{
A list with two items:
\item{d}{A spatial objects data.frame (sf) containing the natural water features within the buffer.}
\item{b}{A spatial objects data.frame (sf) containing the buffer area around the focal point.}
}
\description{
This function extracts natural water features within a given \code{buffer} around a specified \code{tf_id} location.
}
\examples{
\dontrun{
get_water_features(df, tf_id = 25, buffer = 500)
}
}
