% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_BNG_lon_lat.R
\name{os_grid_to_lat_lon}
\alias{os_grid_to_lat_lon}
\title{Converts Ordnance Survey grid reference to latitude and longitude}
\usage{
os_grid_to_lat_lon(grid)
}
\arguments{
\item{grid}{A character vector of an Ordnance Survey grid reference (e.g. 'TL4871')}
}
\value{
A list with the longitude and latitude in decimal degrees.
}
\description{
This function converts an Ordnance Survey grid reference to latitude and longitude.
It requires the reticulate and tidyverse packages to be installed,
as well as the OSGridConverter python package.
}
