% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os_landcover.R
\name{os_ngd_api_call}
\alias{os_ngd_api_call}
\title{os_ngd_api_call}
\usage{
os_ngd_api_call(bbox, key, offset)
}
\arguments{
\item{bbox}{The bounding box as a spatial object or a numeric vector with xmin, ymin, xmax, ymax coordinates.}

\item{key}{The API key for the Ordnance Survey Data Hub.}

\item{offset}{The offset value for pagination.}
}
\value{
A list with two elements:
\itemize{
\item \code{os_lc}: The result of the API call as a spatial object of class \code{sf}.
\item \code{url}: The API call URL used.
}
}
\description{
Calls the Ordnance Survey OpenSpace National Geographic Database API and returns a list containing the results and the API call URL.
}
\examples{
bbox <- c(-2.16, 53.9, -2.13, 53.92)
key <- Sys.getenv('OS_DATA_HUB')
os_ngd_api_call(bbox, key, 0)

}
