% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyForestR-package.R
\docType{package}
\name{tinyfoRest-package}
\alias{tinyfoRest}
\alias{tinyfoRest-package}
\title{tinyfoRest: Get landscape and biodiversity data about Tiny Forests® and surrounding areas}
\description{
Retrieves a range of data about Tiny Forests
}
\author{
\strong{Maintainer}: Julian Flowers \email{julian.flowers12@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
